+++ crates/relayer-framework/src/base/impls/packet_relayers/mod.rs
@@ -1,6 +1,6 @@
 pub mod base_ack_packet;
 pub mod base_receive_packet;
+pub mod base_timeout_unordered_packet;
 pub mod full_relay;
 pub mod retry;
 pub mod skip_received_packet;
-pub mod timeout_unordered_packet;
